requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.2.0
certifi==2020.12.5
chardet==4.0.0
dateparser==1.0.0
idna==2.10
multidict==5.1.0
pyaes==1.6.1
pyasn1==0.4.8
python-binance==1.0.7
python-dateutil==2.8.1
pytz==2021.1
regex==2021.4.4
requests==2.25.1
rsa==4.7.2
six==1.16.0
Telethon==1.21.1
typing-extensions==3.10.0.0
tzlocal==2.1
ujson==4.0.2
urllib3==1.26.4
websockets==9.0.1
wincertstore==0.2
yarl==1.6.3